+++ services/backend/src/utils/bloomFilter.worker.js
@@ -2,12 +2,14 @@
 const config = require('config');
 const { parentPort } = require('worker_threads');
 const { BloomFilter } = require('bloomit');
+const { performance } = require('perf_hooks');
 
 const { SHA256, uuidToHex } = require('@lucaapp/crypto');
 
 const FALSE_POSITIVE_RATE = config.get('bloomFilter.falsePositiveRate');
 
 const unregisteredBadgeBloomFilter = async unregisteredBadges => {
+  const t0 = performance.now();
   const filter = BloomFilter.create(
     unregisteredBadges.length,
     FALSE_POSITIVE_RATE
@@ -16,7 +18,10 @@ const unregisteredBadgeBloomFilter = async unregisteredBadges => {
     const uuidSHA256 = SHA256(uuidToHex(badge.uuid));
     filter.add(uuidSHA256);
   });
-  return filter.export();
+  return {
+    bloomFilterArrayDump: filter.export(),
+    time: performance.now() - t0,
+  };
 };
 
 parentPort.on('message', async unregisteredBadges => {