+++ services/webapp/src/components/TextInput/index.js
@@ -1,1 +1,1 @@
-export * from './TextInput.react.js';
+export * from './TextInput.react';